+++ src/components/new-simulation/button-toggle.tsx
@@ -1,8 +1,7 @@
-import { clsx } from "clsx";
 import React from "react";
-
+import { clsx } from "clsx";
 import { useAppContext } from "../../hooks/use-app-context";
-import { invertedControl } from "../../hooks/use-common-state";
+import { invertedControl } from "../../utils/control-utils";
 
 import "./button-toggle.scss";
